Refactor wp-api converters and utils, and add closeOnOutsideClick function
